refactor!: move Operation to domain/access
