+++ todoapp/application/task/commands/create_task.py
@@ -7,7 +7,7 @@ from todoapp.application.common.command import Command, CommandHandler
 from todoapp.application.common.interfaces.uow import UnitOfWork
 from todoapp.application.task.interfaces import TaskRepo,TaskListGetter
 from todoapp.application.task_list.exceptions import TaskListAccessError
-from todoapp.domain.common.constants import Operation
+from todoapp.domain.access import Operation
 from todoapp.domain.task.entities import Task
 from todoapp.domain.task_list.value_objects import ListId
 from todoapp.domain.user.entities import UserId
