+++ todoapp/application/task/commands/put_in_list.py
@@ -9,7 +9,7 @@ from todoapp.application.task.dto import Task
 from todoapp.application.task.exceptions import TaskAccessError
 from todoapp.application.task.interfaces import TaskRepo,TaskListGetter
 from todoapp.application.task_list.exceptions import TaskListAccessError
-from todoapp.domain.common.constants import Operation
+from todoapp.domain.access import Operation
 from todoapp.domain.task.value_objects import TaskId
 from todoapp.domain.task_list.value_objects import ListId
 from todoapp.domain.user.entities import UserId
