+++ todoapp/application/task_list/commands/delete_task_list.py
@@ -5,7 +5,7 @@ from todoapp.application.common.command import Command, CommandHandler
 from todoapp.application.common.interfaces.uow import UnitOfWork
 from todoapp.application.task_list.exceptions import TaskListAccessError
 from todoapp.application.task_list.interfaces import TaskListRepo
-from todoapp.domain.common.constants import Operation
+from todoapp.domain.access import Operation
 from todoapp.domain.task_list.value_objects import ListId
 from todoapp.domain.user.entities import UserId
