+++ todoapp/domain/common/constants.py
@@ -1,17 +1,7 @@
-from enum import Enum, auto
+from enum import Enum
 
 
 class Empty(Enum):
     UNSET = "UNSET"
 
 
-class Operation(Enum):
-    read = auto()
-    update_list = auto()
-    delete_list = auto()
-    edit_sharing = auto()
-
-    update_task = auto()
-
-    add_task_to_list = auto()
-    delete_task_from_list = auto()
